+++ decidim-core/app/cells/decidim/collapsible_list_cell.rb
@@ -5,9 +5,8 @@ module Decidim
   # `model` array will be rendered with the `:cell` cell.
   #
   # Available sizes:
-  #  - `:small` => collapses after 3 elements.
-  #  - `:default` => collapses after 7 elements. If not specified, this one is
-  #    used.
+  #  - any number between 1 and 12
+  #  - default value is 3
   #
   # Example:
   #
@@ -17,11 +16,9 @@ module Decidim
   #      cell_name: "my/cell",
   #      cell_options: { my: :options },
   #      hidden_elements_count_i18n_key: "my.custom.key",
-  #      size: :small
+  #      size: 4
   #    )
   class CollapsibleListCell < Decidim::ViewModel
-    MIN_LENGTH_FOR_SIZE = { small: 3, default: 7 }.freeze
-
     private
 
     def list
@@ -37,22 +34,20 @@ module Decidim
     end
 
     def size
-      return :small if options[:size].to_s == "small"
-      :default
+      options[:size] || 3
     end
 
     def list_size_class
-      return "small" if size == :small
-      ""
+      "show-#{size}"
     end
 
     def collapsible?
-      list.size > MIN_LENGTH_FOR_SIZE[size]
+      list.size > size
     end
 
     def hidden_elements_count
       return 0 unless collapsible?
-      list.size - MIN_LENGTH_FOR_SIZE[size]
+      list.size - size
     end
 
     def hidden_elements_count_i18n_key